dspy-optimize/requirements.txt
dspy>=2.5.0
anthropic>=0.40.0
python-dotenv>=1.0.0
